% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liteformats-package.R
\docType{package}
\name{liteformats-package}
\alias{liteformats}
\alias{liteformats-package}
\title{Lightweight output formats for litedown}
\description{
Provides HTML-first resume and cover-letter formats with self-contained
styling, local pages.js pagination, and optional PDF output through a
headless Chromium-based browser.
}
\details{
Start with \code{\link[=use_resume]{use_resume()}} or \code{\link[=use_cover_letter]{use_cover_letter()}}, then render with
\code{\link[=resume]{resume()}} or \code{\link[=cover_letter]{cover_letter()}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nanx.me/liteformats/}
  \item \url{https://github.com/nanxstats/liteformats}
  \item Report bugs at \url{https://github.com/nanxstats/liteformats/issues}
}

}
\author{
\strong{Maintainer}: Nan Xiao \email{me@nanx.me} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Nan Xiao \email{me@nanx.me} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID}) [copyright holder]
}

}
\keyword{internal}
