% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liteformats-package.R
\docType{package}
\name{liteformats-package}
\alias{liteformats}
\alias{liteformats-package}
\title{liteformats: Lightweight Output Formats for 'litedown'}
\description{
A collection of lightweight, minimalist output formats and templates for 'litedown', including resumes, cover letters, presentations, and other common document types.
}
\author{
\strong{Maintainer}: Nan Xiao \email{me@nanx.me} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Nan Xiao \email{me@nanx.me} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID}) [copyright holder]
}

}
\keyword{internal}
